\name{acidification_equivalents}

\alias{acidification_equivalents}

\title{
Acidification equivalents
}

\description{
Get a table with acidification equivalents of nitrogen fertilizers. Acidification equivalents are expressed as kg of Effective Calcium Carbonate (ECC) per kg of N.
}



\usage{
acidification_equivalents()
}


\value{
a data.frame with the fertilizer product name, their N content, and the acidification equivalents depending on the fertilizer recovery by the crop. Recover is either zero (R0), 50 (R50) or 100\% (R100).
} 

\examples{
acidification_equivalents()
}
