% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_survey.R
\name{activate_survey}
\alias{activate_survey}
\title{activate_survey}
\usage{
activate_survey(iSurveyID, verbose = FALSE)
}
\arguments{
\item{iSurveyID}{integer, Id of the survey to be activated}

\item{verbose}{boolean, Giving out logging info}
}
\description{
Sets the status of a survey to active
}
\examples{
\dontrun{
activate_survey("475835", verbose = TRUE)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
