% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group.R
\name{add_group}
\alias{add_group}
\title{Add a question group to a survey}
\usage{
add_group(survey_id, properties, quiet = FALSE)
}
\arguments{
\item{survey_id}{Survey id}

\item{properties}{group properties\cr
A data frame is required with at least a column named \code{group_name} (for group name).\cr
Optional columns ca also be added : \code{language} (for group language), \code{description} (for group description) and \code{grelevance} (for group relevance).}

\item{quiet}{If \code{TRUE} then there is no output during group creation}
}
\value{
The created group id.
}
\description{
This function adds a question group to an existing survey using a LSG template file.\cr
Compared to add_group API call, it can add a multilingual question group.
}
\examples{
\dontrun{

limer::add_group(
  survey_id = 123456,
  properties = data.frame(
    group_name = c("Group 1", "Group 2"),
    description = c("Description 1", NA)
  )
)
}

}
