% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_participants.R
\name{add_participants}
\alias{add_participants}
\title{add_participants}
\usage{
add_participants(iSurveyID, data, bCreateToken = F, chunksize = 200)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{data}{dataframe with the columns firstname, lastname and email}

\item{bCreateToken}{boolean Should tokens be created}

\item{chunksize}{integer, size of chunks for handling php memory problems}
}
\value{
API Response
}
\description{
Adds participants to a survey
}
\references{
https://api.limesurvey.org/classes/remotecontrol_handle.html#method_add_participants
}
