% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_question.R
\name{add_question}
\alias{add_question}
\title{Add a question to a survey}
\usage{
add_question(survey_id, group_id, title, type, question, properties,
  subquestions = NULL, answers = NULL, quiet = FALSE)
}
\arguments{
\item{survey_id}{Survey id}

\item{group_id}{Question group id}

\item{title}{Question code}

\item{type}{Question type (see \url{https://manual.limesurvey.org/Expression_Manager#Qcode_Variable_Naming} and table \link{question_types})}

\item{question}{Question text\cr
It can be a single character in case of a single language survey.\cr
Otherwise, a data frame with 2 columns is required : \code{language} (for question language) and \code{question} (for question text).\cr
1 optional column can be added : \code{help} (for question help).}

\item{properties}{Question properties as a named list (see \url{https://api.limesurvey.org/classes/Question.html} and table \link{question_attributes})\cr
Additional question properties are accepted.\cr
A list element can contain multiple values in case of a multilingual attribute (See \code{limer::question_attributes} with attributes whose option equals \code{'i18n'=>true}).}

\item{subquestions}{Subquestions data frame with 4 columns : \code{language} (for subquestion language), \code{title} (for subquestion code), \code{question} (for subquestion text) and \code{relevance} (for conditionnal equation).\cr}

\item{answers}{Answers data frame with 3 columns : \code{language} (for answer language), \code{code} (for answer code) and \code{answer} (for answer text).\cr}

\item{quiet}{If \code{TRUE} then there is no output during question creation}
}
\description{
This function adds a question to an existing survey using a LSQ template file.\cr
This can be a multilingual question (see details).\cr
It also accepts additionnal properties (see table \link{question_attributes}).
}
\details{
If any of the 3 multi-languages arguments is set (eg : \code{text}, \code{subquestions} and \code{answers}) then all these 3 arguments must conform to multi-languages specifications.\cr
Otherwise a error is raised.
}
\examples{
\dontrun{

# Example with long_text
limer::add_question(
  survey_id = 123456,
  group_id = 12345,
  title = "q1",
  type = "T",
  question = data.frame(
    question = "Text question",
    help = "Help 1"
  ),
  properties = data.frame(
    key = "mandatory",
    value = "Y"
  )
)

# Example with array_radio

limer::add_question(
  survey_id = 123456,
  group_id = 12345,
  title = "q3",
  type = "F",
  question = "Array radio question",
  properties = data.frame(
    key = "other",
    value = "Y"
  ),
  subquestions = data.frame(
    title = c("SQ001", "SQ002"),
    question = c("Subquestion 1", "Subquestion 2")
  ),
  answers = data.frame(
    code = c("A1", "A2", "A3"),
    answer = c("Answer 1", "Answer 2", "Answer 3")
  )
)
}

}
