% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_survey.R
\name{add_survey}
\alias{add_survey}
\title{Add a survey}
\usage{
add_survey(sid, properties, attributedescriptions = NULL, quiet = FALSE)
}
\arguments{
\item{sid}{Survey id}

\item{properties}{Survey properties (See table \link{survey_properties})}

\item{attributedescriptions}{Participants attributes defined in a data.frame}

\item{quiet}{If \code{TRUE} then there is no output during survey creation}
}
\value{
The created survey id (can differ from the original if already used)
}
\description{
This function adds a survey using a LSS template file.\cr
Compared to add_survey API call, it allows to automatise the participants attributes creation.
}
\examples{
\dontrun{

limer::add_survey(
  survey_id = 123456,
  properties = data.frame(
    key = c("language", "surveyls_title", "admin", "showprogress"),
    language = c(NA, "en", NA, NA),
    value = c("en", "Survey title", "Survey administrator", "Y")
  ),
  participants_attributes = data.frame(
    description = c("attribute 1 description")
  )
)
}

}
