% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64_to_df.R
\name{base64_to_df}
\alias{base64_to_df}
\title{Convert base64 encoded data to a data frame}
\usage{
base64_to_df(x)
}
\arguments{
\item{x}{\dots}
}
\description{
This function converts raw base64 results into a data frame.
}
\examples{
\dontrun{
base64_to_df()
}
}

