% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{completed_responses}
\alias{completed_responses}
\title{Export data from a LimeSurvey survey}
\usage{
completed_responses(iSurveyID, session = FALSE)
}
\arguments{
\item{iSurveyID}{\dots}

\item{session}{\dots}
}
\description{
Export data from a LimeSurvey survey
}
