% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey.R
\name{create_survey}
\alias{create_survey}
\title{create_survey}
\usage{
create_survey(lss_file, tbl_participants, survey_id = NULL,
  survey_title = NULL, attributes_description = NULL)
}
\arguments{
\item{lss_file}{\dots}

\item{tbl_participants}{\dots}

\item{survey_id}{\dots}

\item{survey_title}{\dots}

\item{attributes_description}{\dots}
}
\description{
create_survey
}
