% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_participants.R
\name{delete_participants}
\alias{delete_participants}
\title{delete_participants}
\usage{
delete_participants(
  iSurveyID,
  aTokenIDs = NULL,
  max_id = 1e+06,
  chunksize = 5000,
  ask = TRUE
)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to delete participants}

\item{aTokenIDs}{integer, Vector with the ID in the participant table}

\item{max_id}{integer, up to which maximum id should the participants be
deleted? only necessary if aTokenIDs is NULL}

\item{chunksize}{integer, size of chunks for handling php memory problems}

\item{ask}{boolean, if TRUE (default) asks for interactive confirmation
before deleting. Set to FALSE for non-interactive/scripted use.}
}
\description{
delete participants from a survey by id
}
\examples{
\dontrun{
delete_participants(475835)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_delete_participants}
}
