% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_responses.R
\name{delete_responses}
\alias{delete_responses}
\title{delete_responses}
\usage{
delete_responses(iSurveyID, iResponseIDs = NULL, verbose = FALSE)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{iResponseIDs}{integer, response Ids to be deleted. If not set delete
all.}

\item{verbose}{boolean, Giving out logging info}
}
\value{
-
}
\description{
Deletes one or more responses
}
\examples{
\dontrun{
delete_responses(475835)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
