% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_gender_column.R
\name{detect_gender_column}
\alias{detect_gender_column}
\title{detect_gender_column}
\usage{
detect_gender_column(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
This helper function corrects an incorrectly to boolean converted column
"gender" back to character.
}
