% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateOrder.R
\name{estimateOrder}
\alias{estimateOrder}
\title{Estimate Order Fees}
\usage{
estimateOrder(account, api.key, symbol, quantity, side, price = 0)
}
\arguments{
\item{account}{The account ID for which to estimate order fees}

\item{api.key}{The API key for authentication and authorization}

\item{symbol}{The trading symbol of the instrument for which to estimate fees}

\item{quantity}{The quantity of the instrument for which to estimate fees}

\item{side}{The side of the order, either 'buy' or 'sell'}

\item{price}{The price at which to estimate fees (default is 0)}
}
\value{
A parsed content of the response indicating the estimated order fees
}
\description{
This function allows you to estimate order fees using the provided account ID and API key.
}
\examples{
account <- "your_account_id"
api.key <- "your_api_key"
symbol <- "AAPL"
quantity <- 10
side <- "buy"
#fee_estimation <- estimateOrder(account, api.key, symbol, quantity, side)
}
