% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists_participants_table.R
\name{exists_participants_table}
\alias{exists_participants_table}
\title{exists_participants_table}
\usage{
exists_participants_table(iSurveyID)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey}
}
\value{
boolean
}
\description{
checks if a participant table already exists
}
\examples{
\dontrun{
exists_participants_table(475835)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
