% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_answer_options.R
\name{get_answer_options}
\alias{get_answer_options}
\title{get_answer_options}
\usage{
get_answer_options(iQuestionID, sUsername = NULL)
}
\arguments{
\item{iQuestionID}{integer, ID of the question
(can be extracted with get_questions_properties)}

\item{sUsername}{character, Username in Limesurvey installation}
}
\value{
Named character vector with answer codes as names and answer text as values
}
\description{
Extracts answer options for a specific question in a LimeSurvey survey
}
