% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes_descriptions.R
\name{get_attributes_descriptions}
\alias{get_attributes_descriptions}
\title{get_attributes_descriptions}
\usage{
get_attributes_descriptions(iSurveyID)
}
\arguments{
\item{iSurveyID}{\dots}
}
\description{
get_attributes_descriptions
}
