% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kontaktdatentabelle.R
\name{get_kontaktdatentabelle}
\alias{get_kontaktdatentabelle}
\title{Fetch a Table from Grist}
\usage{
get_kontaktdatentabelle(
  table_name = "Zentrale_datenbank",
  doc_id = "r2HXjB5zhKA8LiuzxQYxmd",
  server = "https://grist.swr-datalab.de",
  api_key = Sys.getenv("GRIST_API_KEY")
)
}
\arguments{
\item{table_name}{A character string specifying the Grist table name.
The first letter needs to be capitalized.
Defaults to \code{"Zentrale_datenbank"}.}

\item{doc_id}{A character string specifying the Grist document ID.
Defaults to \code{"r2HXjB5zhKA8"}.}

\item{server}{A character string specifying the Grist server URL.
Defaults to \code{"https://grist.swr-datalab.de"}.}

\item{api_key}{A character string containing the Grist API key.
Defaults to \code{GRIST_API_KEY} from the calling environment.}
}
\value{
A data frame of records from the specified Grist table.
}
\description{
Retrieves records from a Grist document table and returns them as a
clean data frame with the \code{fields.} prefix removed from column names.
}
