% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_property.R
\name{get_participant_property}
\alias{get_participant_property}
\title{Get a participant property from a LimeSurvey survey}
\usage{
get_participant_property(
  iSurveyID,
  aTokenQueryProperties,
  aTokenProperties = NULL
)
}
\arguments{
\item{iSurveyID}{\dots}

\item{aTokenQueryProperties}{\dots}

\item{aTokenProperties}{\dots}
}
\description{
This function exports and downloads a participant property from a LimeSurvey
survey.
}
\examples{
\dontrun{
get_participant_property(iSurveyID = 475835,
                                   aTokenQueryProperties = 1)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
