% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participants.R
\name{get_participants}
\alias{get_participants}
\title{get_participants}
\usage{
get_participants(
  iSurveyID,
  bUnused = TRUE,
  iStart = 1,
  iLimit = 100,
  tid = FALSE,
  chunksize = 5000,
  ...
)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to retrieve participants from}

\item{bUnused}{boolean, if TRUE, only unused tokens are returned}

\item{iStart}{integer, start id of the token list}

\item{iLimit}{integer, number of participants to return}

\item{tid}{boolean, if TRUE, includes the tid column in the result}

\item{chunksize}{integer, size of chunks used to split large requests
and avoid php memory problems}

\item{...}{ellipsis parameters passed on to call_limer}
}
\value{
dataframe of participant data
}
\description{
Retrieves the list of participants of a survey. Large requests are
automatically split into chunks to avoid PHP memory issues on the
LimeSurvey server.
}
\examples{
\dontrun{
get_participants(475835)
}
}
\references{
https://api.limesurvey.org/classes/remotecontrol_handle.html#method_list_participants
}
