% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{get_responses}
\title{Export data from a LimeSurvey survey}
\usage{
get_responses(
  iSurveyID,
  sDocumentType = "csv",
  sLanguageCode = NULL,
  sCompletionStatus = "complete",
  sHeadingType = "code",
  sResponseType = "long",
  ...
)
}
\arguments{
\item{iSurveyID}{string ID of the Survey}

\item{sDocumentType}{string any format available by plugins (for example :
pdf, csv, xls, doc, json).}

\item{sLanguageCode}{string (optional) The language to be used.}

\item{sCompletionStatus}{string (optional) 'complete','incomplete' or 'all'
- defaults to 'all'.}

\item{sHeadingType}{string (optional) 'code','full' or 'abbreviated'.
defaults to 'code'.}

\item{sResponseType}{(optional)'short' or 'long'. defaults to 'short'}

\item{\dots}{Further arguments passed to internal function \code{call_limer()}.}
}
\description{
This function exports and downloads data from a LimeSurvey survey.
}
\examples{
\dontrun{
get_responses(475835)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
