% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_key.R
\name{get_session_key}
\alias{get_session_key}
\title{Get a LimeSurvey API session key}
\usage{
get_session_key(
  username = getOption("lime_username"),
  password = getOption("lime_password"),
  lime_api = getOption("lime_api"),
  ssl_verifypeer = FALSE
)
}
\arguments{
\item{username}{LimeSurvey username. Defaults to value set in \code{options()}.}

\item{password}{LimeSurvey password Defaults to value set in \code{options()}.}

\item{lime_api}{string, URL of the LimeSurvey RemoteControl API endpoint.
Defaults to the value set via \code{options(lime_api = ...)}; if unset,
falls back to the SWR Medienanfragen installation's endpoint.}

\item{ssl_verifypeer}{boolean \code{httr::config()} parameter. Default is
FALSE.}
}
\value{
API token
}
\description{
This function logs into the LimeSurvey API and provides an access session key.
}
\examples{
\dontrun{
get_session_key()
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_session_key}
}
