% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uploaded_files.R
\name{get_uploaded_files}
\alias{get_uploaded_files}
\title{Obtain all uploaded files for all responses}
\usage{
get_uploaded_files(iSurveyID, responseId, sToken = "", downloadpath = ".")
}
\arguments{
\item{iSurveyID}{\dots}

\item{iStart}{\dots}

\item{iLimit}{\dots}

\item{bUnused}{\dots}

\item{aAttributes}{\dots}
}
\description{
This function downloads all uploaded files to all responses from a LimeSurvey survey.
}
\examples{
\dontrun{
get_uploaded_files(12345)
get_uploaded_files(12345, iStart=1, iLimit=10, bUnused=FALSE, aAttributes=FALSE)
}
}
\references{
https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_uploaded_files
}
