% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccount.R
\name{getAccount}
\alias{getAccount}
\title{Get Account Information}
\usage{
getAccount(api.key)
}
\arguments{
\item{api.key}{The API key for authentication and authorization}
}
\value{
A data frame containing account information
}
\description{
This function retrieves account information using the provided API key.
}
\examples{
api.key <- "your_api_key"
#account_info <- getAccount(api.key)

}
