% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistory.R
\name{getHistory}
\alias{getHistory}
\title{Get Historical Market Data}
\usage{
getHistory(
  symbol,
  api.key,
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  period = "day"
)
}
\arguments{
\item{symbol}{A character vector of the symbol for which to retrieve historical data}

\item{api.key}{The API key for authentication and authorization}

\item{start_date}{The start date for the historical data (default is 30 days ago)}

\item{end_date}{The end date for the historical data (default is the current date)}

\item{period}{The period for the historical data (default is 'day')}
}
\value{
A parsed content of the response containing the historical market data
}
\description{
This function retrieves historical market data for a specific symbol using the provided API key.
}
\examples{
symbol <- "AAPL"
api.key <- "your_api_key"
start_date <- "2023-01-01"
end_date <- "2023-02-01"
period <- "day"
#historical_data <- getHistory(symbol, api.key, start_date, end_date, period)
#getHistory('AAPL',apikey)
}
