% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrders.R
\name{getOrders}
\alias{getOrders}
\title{Get Active Orders}
\usage{
getOrders(account, api.key)
}
\arguments{
\item{account}{The account ID for which to retrieve active orders}

\item{api.key}{The API key for authentication and authorization}
}
\value{
A parsed content of the response containing the active orders for the account
}
\description{
This function retrieves active orders for the provided account using the provided API key.
}
\examples{
account <- "your_account_id"
api.key <- "your_api_key"
active_orders <- getOrders(account, api.key)
#getOrders('dmo00374@demo',apikey)

}
