% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrderStatus.R
\name{getOrderStatus}
\alias{getOrderStatus}
\title{Get Order Status}
\usage{
getOrderStatus(api.key, order_id)
}
\arguments{
\item{api.key}{The API key for authentication and authorization}

\item{order_id}{The ID of the order for which to retrieve the status}
}
\value{
A parsed content of the response indicating the status of the order
}
\description{
This function retrieves the status of a specific order using the provided API key and order ID.
}
\examples{
api.key <- "your_api_key"
order_id <- "your_order_id"
order_status <- getOrderStatus(api.key, order_id)
#getOrderStatus(apikey,234684792874)

}
