% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPositions.R
\name{getPositions}
\alias{getPositions}
\title{Get Account Positions}
\usage{
getPositions(account, api.key, date = Sys.Date())
}
\arguments{
\item{account}{The account ID for which to retrieve positions}

\item{api.key}{The API key for authentication and authorization}

\item{date}{The date for which to retrieve positions (default is current date)}
}
\value{
A data frame containing account positions
}
\description{
This function retrieves account positions using the provided account ID and API key.
}
\examples{
account <- "your_account_id"
api.key <- "your_api_key"
#positions <- getPositions(account, api.key)
#getPositions('dmo00374@demo',apikey)
#getPositions('dmo00374@demo',apikey,date=Sys.Date()-4)

}
