% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{Get Quote for Symbols}
\usage{
getQuote(symbol, api.key)
}
\arguments{
\item{symbol}{A character vector of symbols for which to retrieve quotes}

\item{api.key}{The API key for authentication and authorization}
}
\value{
A parsed content of the response containing the quotes for the provided symbols
}
\description{
This function retrieves quotes for multiple symbols using the provided API key.
}
\examples{
symbols <- c("AAPL", "MSFT")
api.key <- "your_api_key"
#quotes <- getQuote(symbols, api.key)
#getQuote('NVDA',apikey)

}
