% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSchedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Get Market Schedule}
\usage{
getSchedule(api.key)
}
\arguments{
\item{api.key}{The API key for authentication and authorization}
}
\value{
A parsed content of the response containing the market schedule
}
\description{
This function retrieves the market schedule using the provided API key.
}
\examples{
api.key <- "your_api_key"
#schedule <- getSchedule(api.key)

}
