% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Get Access Token}
\usage{
getToken(client_id, client_secret, username, password)
}
\arguments{
\item{client_id}{The client ID provided by Lime}

\item{client_secret}{The client secret provided by Lime}

\item{username}{Your Lime account username}

\item{password}{Your Lime account password}
}
\value{
The access token required for accessing Lime APIs
}
\description{
This function retrieves an access token from Lime's authentication server
}
\examples{
client_id <- "your_client_id"
client_secret <- "your_client_secret"
username <- "your_username"
password <- "your_password"
#access_token <- getToken(client_id, client_secret, username, password)

}
