% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrades.R
\name{getTrades}
\alias{getTrades}
\title{Get Account Trades}
\usage{
getTrades(account, api.key, date = Sys.Date(), limit = 10, skip = 0)
}
\arguments{
\item{account}{The account ID for which to retrieve trades}

\item{api.key}{The API key for authentication and authorization}

\item{date}{The date for which to retrieve trades (default is current date)}

\item{limit}{The maximum number of trades to retrieve (default is 10)}

\item{skip}{The number of trades to skip (default is 0)}
}
\value{
A data frame containing account trades
}
\description{
This function retrieves account trades using the provided account ID and API key.
}
\examples{
account <- "your_account_id"
api.key <- "your_api_key"
#trades <- getTrades(account, api.key)
#getTrades('dmo00374@demo',apikey)
#getTrades('dmo00374@demo',apikey,date=Sys.Date()-2,limit=2,skip=30)

}
