% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransactions.R
\name{getTransactions}
\alias{getTransactions}
\title{Get Account Transactions}
\usage{
getTransactions(
  account,
  api.key,
  start_date = Sys.Date(),
  end_date = Sys.Date(),
  limit = 10,
  skip = 0
)
}
\arguments{
\item{account}{The account ID for which to retrieve transactions}

\item{api.key}{The API key for authentication and authorization}

\item{start_date}{The start date for retrieving transactions (default is current date)}

\item{end_date}{The end date for retrieving transactions (default is current date)}

\item{limit}{The maximum number of transactions to retrieve (default is 10)}

\item{skip}{The number of transactions to skip (default is 0)}
}
\value{
A list of parsed JSON containing account transactions
}
\description{
This function retrieves account transactions using the provided account ID and API key.
}
\examples{
account <- "your_account_id"
api.key <- "your_api_key"
#transactions <- getTransactions(account, api.key)
#getTransactions('dmo00374@demo',apikey)
#getTransactions('dmo00374@demo',apikey,limit=25,skip=0,end_date='2023-08-22')

}
