% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participants.R
\name{list_participants}
\alias{list_participants}
\title{list_participants}
\usage{
list_participants(...)
}
\arguments{
\item{...}{ellipsis parameters passed on to call_limer}
}
\description{
Deprecated alias for [get_participants()]. Use that function instead.
}
