% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_registered_participant.R
\name{mail_registered_participant}
\alias{mail_registered_participant}
\title{mail_registered_participant}
\usage{
mail_registered_participant(iSurveyID, tid)
}
\arguments{
\item{iSurveyID}{\dots}

\item{tid}{\dots}
}
\description{
mail_registered_participant
}
\examples{
\dontrun{
mail_registred_participant(iSurveyID = 123456, tid = 2)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
