% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailing.R
\name{mailing_create_survey}
\alias{mailing_create_survey}
\title{Mailing}
\usage{
mailing_create_survey(from, to, subject, body, session = FALSE)
}
\arguments{
\item{from}{Expéditeur list("email", "alias")}

\item{to}{Tibble des destinataires}

\item{subject}{Sujet du mail}

\item{body}{Corps du mail}

\item{session}{Open a limesurvey session}
}
\description{
Mailing
}
