% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionChain.R
\name{optionChain}
\alias{optionChain}
\title{Option Chain Lookup}
\usage{
optionChain(symbol, series, api.key)
}
\arguments{
\item{symbol}{A character vector representing the symbol for which to retrieve the option chain}

\item{series}{A character vector representing the option series (expiration) for which to retrieve the chain}

\item{api.key}{The API key for authentication and authorization}
}
\value{
A parsed content of the response containing the option chain information
}
\description{
This function retrieves an option chain for a specific symbol and series (expiration) using the provided API key.
}
\examples{
symbol <- "AAPL"
series <- "2023-08-25"
api.key <- "your_api_key"
#option_chain <- optionChain(symbol, series, api.key)
#optionChain('AAPL','2023-09-01',apikey)
}
