% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionSeries.R
\name{optionSeries}
\alias{optionSeries}
\title{Option Series Lookup}
\usage{
optionSeries(symbol, api.key)
}
\arguments{
\item{symbol}{A character vector representing the symbol for which to retrieve option series}

\item{api.key}{The API key for authentication and authorization}
}
\value{
A parsed content of the response containing the option series information
}
\description{
This function retrieves option series for a specific symbol using the provided API key.
}
\examples{
symbol <- "AAPL"
api.key <- "your_api_key"
#option_series <- optionSeries(symbol, api.key)
#optionSeries('AAPL',apikey)

}
