% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling `rhs(lhs)`.
}
\description{
`limer` uses the magrittr pipe throughout its functions (e.g. chaining
`dplyr::filter()`/`dplyr::pull()` calls on survey data). This operator
is re-exported so it's available to users of the package without
requiring a separate `library(magrittr)` call.
}
\details{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
