% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{question_attributes}
\alias{question_attributes}
\title{Question attributes table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 137 rows and 10 columns.}
\usage{
question_attributes
}
\description{
It is based on this web page \url{https://github.com/LimeSurvey/LimeSurvey/blob/master/application/helpers/questionHelper.php}.
}
\keyword{datasets}
