% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{question_types}
\alias{question_types}
\title{Question types table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 29 rows and 3 columns.}
\usage{
question_types
}
\description{
Question types table
}
\keyword{datasets}
