% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_session_key.R
\name{release_session_key}
\alias{release_session_key}
\title{Release a LimeSurvey API session key}
\usage{
release_session_key()
}
\description{
This function clears the LimeSurvey API session key currently in use, effectively logging out.
}
\examples{
\dontrun{
release_session_key()
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
