% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_string.R
\name{sanitize_string}
\alias{sanitize_string}
\title{sanitize_string}
\usage{
sanitize_string(input_vector)
}
\arguments{
\item{input_vector}{string}
}
\value{
sanitized string
}
\description{
A function to sanitize a string and make it a valid name
}
