% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_participant_property.R
\name{set_participant_property}
\alias{set_participant_property}
\title{Set a participant property from a LimeSurvey survey}
\usage{
set_participant_property(iSurveyID, aTokenQueryProperties, aTokenData)
}
\arguments{
\item{iSurveyID}{\dots}

\item{aTokenQueryProperties}{\dots}

\item{aTokenProperties}{\dots}
}
\description{
This function sets a participant property from a LimeSurvey survey.
}
\examples{
\dontrun{
set_participant_properties(iSurveyID = 12345, aTokenQueryProperties = 1, aTokenData = list("attribute_1" = "test"))
}
}
