% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_participants_completed}
\alias{set_participants_completed}
\title{Set to completed in the particpants table completed responses}
\usage{
set_participants_completed(iSurveyID)
}
\arguments{
\item{iSurveyID}{Survey id}
}
\description{
Set to completed in the particpants table completed responses
}
