% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_base64.R
\name{str_to_base64}
\alias{str_to_base64}
\title{Convert string to base64 encoded data}
\usage{
str_to_base64(str)
}
\arguments{
\item{str}{\dots}
}
\description{
This function converts a string into a raw base64 result.
}
\examples{
\dontrun{
}
}
