% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_properties}
\alias{survey_properties}
\title{Survey properties table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 98 rows and 3 columns.}
\usage{
survey_properties
}
\description{
Survey properties table
}
\keyword{datasets}
