% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolLookUp.R
\name{symbolLookUp}
\alias{symbolLookUp}
\title{Symbol Look-Up}
\usage{
symbolLookUp(symbol, api.key, limit = 5)
}
\arguments{
\item{symbol}{A character vector representing the symbol to look up}

\item{api.key}{The API key for authentication and authorization}

\item{limit}{The maximum number of results to return (default is 5)}
}
\value{
A parsed content of the response containing the symbol look-up results
}
\description{
This function performs a symbol look-up using the provided API key and symbol query.
}
\examples{
symbol <- "AAPL"
api.key <- "your_api_key"
limit <- 10
#symbol_results <- symbolLookUp(symbol, api.key, limit)
#symbolLookUp('AAPL',api.key=apikey)
}
