% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_participants.R
\name{update_participants}
\alias{update_participants}
\title{update_participants}
\usage{
update_participants(id_survey, tbl_update, message = TRUE, sleep = 1)
}
\arguments{
\item{id_survey}{\dots}

\item{tbl_update}{\dots}

\item{message}{\dots}

\item{sleep}{\dots}
}
\description{
update_participants
}
