% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_responses.R
\name{update_responses}
\alias{update_responses}
\title{update_responses}
\usage{
update_responses(survey_id, token, question, value, session = FALSE)
}
\arguments{
\item{survey_id}{\dots}

\item{token}{\dots}

\item{question}{\dots}

\item{value}{\dots}

\item{session}{\dots}
}
\description{
update_responses
}
\examples{

Simple question
sidXgidXqid

Array question
sidXgidXqidTitle

}
