% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateOrder.R
\name{validateOrder}
\alias{validateOrder}
\title{Validate an Order}
\usage{
validateOrder(
  account,
  api.key,
  symbol,
  quantity,
  price = 0,
  stop_price = 0,
  time_in_force = "day",
  order_type = "market",
  side = "buy",
  comment = "",
  exchange = "Auto"
)
}
\arguments{
\item{account}{The account ID for which to validate the order}

\item{api.key}{The API key for authentication and authorization}

\item{symbol}{The trading symbol of the instrument to validate}

\item{quantity}{The quantity of the instrument to validate}

\item{price}{The price at which to validate the order (default is 0)}

\item{stop_price}{The stop price for stop orders (default is 0)}

\item{time_in_force}{The time in force for the order (default is 'day')}

\item{order_type}{The order type (default is 'market')}

\item{side}{The side of the order, either 'buy' or 'sell' (default is 'buy')}

\item{comment}{Additional comment or note for the order (default is '')}

\item{exchange}{The exchange to validate the order on (default is 'Auto')}
}
\value{
A parsed content of the response indicating the success of the order validation
}
\description{
This function allows you to validate an order before placing it, using the provided account ID and API key.
}
\examples{
account <- "your_account_id"
api.key <- "your_api_key"
symbol <- "AAPL"
quantity <- 10
#validation_response <- validateOrder(account, api.key, symbol, quantity)
}
