\name{acidification_equivalents}

\alias{acidification_equivalents}

\title{
Acidification equivalents
}

\description{
Get a table with acidification equivalents of nitrogen fertilizers. Acidification equivalents are expressed as kg of Effective Calcium Carbonate (ECC) per kg of N.
}


\usage{
acidification_equivalents()
}
 

\examples{
acidification_equivalents()
}
