% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_tokens.R
\name{activate_tokens}
\alias{activate_tokens}
\alias{create_participants_table}
\title{activate_tokens}
\usage{
activate_tokens(iSurveyID, aAttributeFields = NULL)

create_participants_table(iSurveyID, aAttributeFields = NULL)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey where a survey participants table
will be created for}

\item{aAttributeFields}{list, An array of integer describing any additional
attribute fields}
}
\value{
status
}
\description{
Initialise the survey participant table of a survey where new participant
tokens may be later added.
}
