% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_responses.R
\name{add_responses}
\alias{add_responses}
\title{add_responses}
\usage{
add_responses(iSurveyID, data, verbose = FALSE)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{data}{dataframe, The actual response(s)}

\item{verbose}{boolean, Giving out logging info}
}
\description{
Inserts one or more answers into an answer table of a survey
}
\references{
https://api.limesurvey.org/classes/remotecontrol_handle.html#method_add_response
}
