% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_limer.R
\name{authenticate_limer}
\alias{authenticate_limer}
\title{authenticate_limer}
\usage{
authenticate_limer(ssl_verifypeer = FALSE)
}
\arguments{
\item{ssl_verifypeer}{boolean, httr::config parameter for API call}
}
\value{
message
}
\description{
Authenticates a user session to a Limesurvey instance. The function
requires that the parameters LIME_USERNAME, LIME_PASSOWRD and
LIME_API_URL are set in the system environment in .Renviron.
}
