% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_survey.R
\name{delete_survey}
\alias{delete_survey}
\title{delete_survey}
\usage{
delete_survey(iSurveyID, verbose)
}
\arguments{
\item{iSurveyID}{integer, Id of the survey to be deleted}

\item{verbose}{boolean, Giving out logging info}
}
\description{
Deletes a survey and all its data
}
\references{
https://api.limesurvey.org/classes/Survey.html#method_deleteSurvey
}
