% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_survey_structure.R
\name{export_survey_structure}
\alias{export_survey_structure}
\title{Export LSS File from a LimeSurvey survey}
\usage{
export_survey_structure(iSurveyID, filename = NULL, verbose = FALSE)
}
\arguments{
\item{iSurveyID}{\dots}

\item{filename}{\dots}

\item{verbose}{boolean, Giving out logging info}
}
\description{
This function exports a survey structure (LSS).
}
\note{
This function requires that the patch has been implemented in the
remotecontrol_handle.php file of your Limesurvey installation.
application/helpers/remotecontrol/remotecontrol_handle.php
system.file("patch.php", package="limer")
See @references
}
\examples{
\dontrun{
export_survey_structure(12345)
}
}
\references{
https://bugs.limesurvey.org/view.php?id=17747
}
