% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participants.R
\name{get_participants}
\alias{get_participants}
\title{Export list of participants from a LimeSurvey survey}
\usage{
get_participants(iSurveyID, iStart, iLimit, bUnused, aAttributes)
}
\description{
This function exports and downloads the list of participants from a LimeSurvey survey.
}
\examples{
\dontrun{
get_responses(12345)
}
}

