% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participants.R
\name{get_participants}
\alias{get_participants}
\title{Export list of participants from a LimeSurvey survey}
\usage{
get_participants(iSurveyID, iStart, iLimit, bUnused, aAttributes)
}
\arguments{
\item{iSurveyID}{\dots}

\item{iStart}{\dots}

\item{iLimit}{\dots}

\item{bUnused}{\dots}

\item{aAttributes}{\dots}
}
\description{
This function exports and downloads the list of participants from a LimeSurvey survey.
}
\examples{
\dontrun{
get_participants(12345, iStart=1, iLimit=10, bUnused=FALSE,
                                   aAttributes=c('attribute_1','attribute_2'))
get_participants(12345, iStart=1, iLimit=10, bUnused=FALSE, aAttributes=FALSE)
}
}

