% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participants.R
\name{get_participants}
\alias{get_participants}
\title{Export list of participants from a LimeSurvey survey}
\usage{
get_participants(
  iSurveyID,
  attributes_descriptions = NULL,
  attributes = NULL,
  all_attributes = FALSE,
  conditions = list(),
  session = FALSE
)
}
\arguments{
\item{iSurveyID}{\dots}

\item{attributes_descriptions}{\dots}

\item{attributes}{\dots}

\item{all_attributes}{\dots}

\item{conditions}{\dots}

\item{session}{\dots}
}
\description{
This function exports and downloads the list of participants from a LimeSurvey survey.
}
