% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions_properties.R
\name{get_questions_properties}
\alias{get_questions_properties}
\title{get_questions_properties}
\usage{
get_questions_properties(iSurveyID, sUsername = NULL, verbose = FALSE)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{sUsername}{character, Username in Limesurvey installation}

\item{verbose}{boolean, Giving out logging info}
}
\value{
dataframe
}
\description{
Generates a dataframe with the properties of all questions in a survey
}
