% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_table_columns.R
\name{get_response_table_columns}
\alias{get_response_table_columns}
\title{get_response_table_columns}
\usage{
get_response_table_columns(iSurveyID, verbose = FALSE, sHeadingType = "full")
}
\arguments{
\item{iSurveyID}{integer, Id of the survey}

\item{verbose}{boolean, Giving out logging info}

\item{sHeadingType}{character, "full" returns the complete text of the
questions, "complete" returns the code. Default = "full"}
}
\value{
String vector
}
\description{
Returns the column names of the response table
}
\note{
This function requires that the patch has been implemented in the
remotecontrol_handle.php file of your Limesurvey installation.
application/helpers/remotecontrol/remotecontrol_handle.php
system.file("patch.php", package="limer")
}
