% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{get_responses}
\title{Export data from a LimeSurvey survey}
\usage{
get_responses(iSurveyID, sDocumentType = "csv", sLanguageCode = "en",
  sCompletionStatus = "complete", sHeadingType = "code",
  sResponseType = "long", ...)
}
\description{
This function exports and downloads data from a LimeSurvey survey.
}
\examples{
\dontrun{
get_responses(12345)
}
}

