% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_key.R
\name{get_session_key}
\alias{get_session_key}
\title{Get a LimeSurvey API session key}
\usage{
get_session_key(username = getOption("lime_username"),
  password = getOption("lime_password"))
}
\arguments{
\item{username}{LimeSurvey username. Defaults to value set in \code{options()}.}

\item{password}{LimeSurvey password Defaults to value set in \code{options()}.}
}
\value{
API token
}
\description{
This function logs into the LimeSurvey API and provides an access session key.
}
\examples{
\dontrun{
get_session_key()
}
}

