% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Get the summary of a LimeSurvey survey}
\usage{
get_summary(iSurveyID, sStatName)
}
\arguments{
\item{iSurveyID}{\dots}

\item{sStatName}{\dots}
}
\description{
This function gives the summary of a LimeSurvey survey.
}
\examples{
\dontrun{
get_summary(iSurveyID = 12345, sStatName = "completed_responses")
}
}
