% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_list.R
\name{get_survey_list}
\alias{get_survey_list}
\title{get_survey_list}
\usage{
get_survey_list(sUsername = NULL, names = FALSE, sid = TRUE)
}
\arguments{
\item{sUsername}{string, username}

\item{names}{boolean, Shall be returned only the names of the surveys.}

\item{sid}{boolean, Shall be given back only the Ids of the surveys.}
}
\value{
dataframe, survey titles (names), or survey ids (sid)
}
\description{
If user is admin he can get surveys of every user (parameter sUsername) or
all surveys (sUsername=null).
}
\references{
https://api.limesurvey.org/classes/remotecontrol_handle.html#method_list_surveys
}
