% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_survey_structure.R
\name{import_survey_structure}
\alias{import_survey_structure}
\title{import_survey_structure}
\usage{
import_survey_structure(
  sImportData,
  sNewSurveyName = NULL,
  DestSurveyID = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sImportData}{string, path to file}

\item{sNewSurveyName}{string, The optional new name of the survey}

\item{DestSurveyID}{integer This is the new ID of the survey - if already
used a random one will be taken instead}

\item{verbose}{boolean, Giving out logging info}
}
\description{
Import survey file in a known format into Limesurvey and create an inactive
survey with it.
}
\references{
https://api.limesurvey.org/classes/remotecontrol_handle.html#method_import_survey
}
