% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limer-package.R
\docType{package}
\name{limer-package}
\alias{limer-package}
\alias{limer}
\title{limer-package}
\description{
A LimeSurvey R Client
}
\details{
Provides access to LimeSurvey's RemoteControl 2 API, allowing
   you to collect and analyze survey data in a simple,
   reproducible workflow.
}
\author{
Andrew Heiss <andrew@andrewheiss.com>
}
\keyword{package}
