% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_participants.R
\name{list_participants}
\alias{list_participants}
\title{list_participants}
\usage{
list_participants(
  iSurveyID,
  bUnused = TRUE,
  iStart = 1,
  iLimit = 100,
  tid = FALSE,
  chunksize = 5000,
  ...
)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{bUnused}{boolean If you want unused tokens, set true}

\item{iStart}{integer, Start id of the token list}

\item{iLimit}{integer, Number of participants to return}

\item{tid}{boolean, return data with tid}

\item{chunksize}{integer, size of chunks for handling php memory problems}

\item{...}{ellipsis parameters for call_limer}
}
\value{
dataframe
}
\description{
Retrieves the list of participants of a survey
}
