% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailing.R
\name{mailing}
\alias{mailing}
\title{Perform mailing}
\usage{
mailing(
  survey_id,
  tid,
  sleep = 7,
  delete = FALSE,
  progress_bar = TRUE,
  session = FALSE
)
}
\arguments{
\item{survey_id}{Id de l'enquête créée}

\item{tid}{Participants tid}

\item{sleep}{Temporisation entre chaque envoi de mail}

\item{delete}{Delete mailing survey after mailing is finished}

\item{progress_bar}{Display a progres bar}

\item{session}{Open a limesurvey session}
}
\description{
Perform mailing
}
