% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mailing.R
\name{mailing}
\alias{mailing}
\title{Mailing}
\usage{
mailing(from, to, subject, body, sleep = 10, delete = FALSE,
  mailing = TRUE)
}
\arguments{
\item{from}{Expéditeur list("email", "alias")}

\item{to}{Tibble des destinataires}

\item{subject}{Sujet du mail}

\item{body}{Corps du mail}

\item{sleep}{Temporisation entre chaque envoi de mail}

\item{delete}{Delete mailing survey after mailing is finished}

\item{mailing}{Delete mailing survey after mailing is finished}
}
\description{
Mailing
}
