% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_participant_properties.R
\name{set_participant_properties}
\alias{set_participant_properties}
\title{set_participant_properties}
\usage{
set_participant_properties(survey_id, tid, property, value)
}
\arguments{
\item{tid}{\dots}

\item{property}{\dots}

\item{value}{\dots}

\item{iSurveyID}{\dots}
}
\description{
set_participant_properties
}
