% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_participant_properties.R
\name{set_participant_properties}
\alias{set_participant_properties}
\title{set_participant_properties}
\usage{
set_participant_properties(iSurveyID, aTokenQueryProperties, aTokenData)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{aTokenQueryProperties}{integer, tid of the user}

\item{aTokenData}{list, key-value-pair of attribute name an value}
}
\value{
none
}
\description{
Allow to set properties about a specific participant, only one
participant can be updated.
}
