% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_participant_properties.R
\name{set_participant_properties}
\alias{set_participant_properties}
\title{set_participant_properties}
\usage{
set_participant_properties(
  iSurveyID,
  aTokenQueryProperties,
  aTokenData,
  verbose = FALSE
)
}
\arguments{
\item{iSurveyID}{integer, ID of the Survey to insert responses}

\item{aTokenQueryProperties}{integer, tid of the user}

\item{aTokenData}{list, key-value-pair of attribute name an value}

\item{verbose}{boolean, Giving out logging info}
}
\value{
none
}
\description{
Allow to set properties about a specific participant, only one
participant can be updated.
}
\examples{
\dontrun{
set_participant_properties(iSurveyID = 475835,
                                     aTokenQueryProperties = list(tid = 1),
                                     aTokenData = list(email = "max.neu@aol.de",
                                                       completed = "Y"),
                                     verbose = TRUE)
}
}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_activate_survey}
}
